+++ src/index-lite.js
@@ -3,28 +3,10 @@
  */
 
 /* eslint-disable import/first */
-if (global.Layer) throw new Error('You appear to have multiple copies of the Layer Web XDK loaded at the same time');
-
-import Constants from './constants';
-import Core from './core';
-import Utils from './utils';
+import { Core, Utils, Constants, init, onInit, version, client, Settings } from './index-core';
 import UI from './ui/index-lite';
-import version from './version';
-import Settings from './settings';
 
-Settings.client = new Core.Client({});
-function init(options) {
-  let client = Settings.client;
-  if (!client || client.isDestroyed) client = Settings.client = new Core.Client({});
-  Object.keys(options).forEach((name) => {
-    Settings[name] = options[name];
-    if (client[name] !== undefined) client[name] = options[name];
-  });
+onInit(UI.init, UI);
 
-  UI.init();
-  return client;
-}
-
-module.exports = { UI, Core, Utils, Constants, init, version, get client() { return Settings.client; }, Settings };
+module.exports = { UI, Core, Utils, Constants, init, onInit, version, client, Settings };
 if (typeof global !== 'undefined') global.Layer = global.layer = module.exports;
-